# Use `AncestralClass` to speed up `Object::cast_to` when possible.