- Forward-port of https://github.com/godotengine/godot/pull/107844
- Follow-up of https://github.com/godotengine/godot/pull/107462

This implementation is 2x faster (hot access, probably more on cold access) than regular `Object::cast_to`, for classes with corresponding ancestral classes.

It would be possible to "pre-check" the types that derive from any of the ancestral types (e.g. `object && object.has_ancestry(AncestralClass::SPATIAL) && object.is_class_ptr(SpatialSubclass::get_class_ptr_static())`). But that would only accelerate the unhappy path, and only some of the time. The happy path wants to immediately use `is_class_ptr`, so I decided against this micro optimization.

I also made `has_ancestry` a private function. Callers should generally use `derives_from` / `Object::cast_to` instead, which is more resistant to us changing the ancestral classes in the future.

## Benchmark

The benchmarks show a ~2x improvement for cold access. The improvement is likely to be better in reality because we generate more hot situations, and need less cold RAM accesses even in cold situations.

Note that the difference is small in absolute terms, due to `cast_to` already being quite well optimized.

```c++
Object *nav_agent = memnew(NavigationAgent3D);
	Object *node_3D = memnew(Node3D);

	{
		auto t0 = std::chrono::high_resolution_clock::now();
		for (int i = 0; i < 1000000000; i++) {
			// Test happy and optimized
			get_ptr_node3d(node_3D);
		}
		auto t1 = std::chrono::high_resolution_clock::now();
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
	{
		auto t0 = std::chrono::high_resolution_clock::now();
		for (int i = 0; i < 1000000000; i++) {
			// Test happy but unoptimized
			get_ptr_navagent(nav_agent);
		}
		auto t1 = std::chrono::high_resolution_clock::now();
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
	{
		auto t0 = std::chrono::high_resolution_clock::now();
		for (int i = 0; i < 1000000000; i++) {
			// Test unhappy but optimized
			get_ptr_node3d(nav_agent);
		}
		auto t1 = std::chrono::high_resolution_clock::now();
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
	{
		auto t0 = std::chrono::high_resolution_clock::now();
		for (int i = 0; i < 1000000000; i++) {
			// Test unhappy and unoptimized
			get_ptr_navagent(node_3D);
		}
		auto t1 = std::chrono::high_resolution_clock::now();
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
```

With the helper functions being declared in a separate module (to avoid inlining):

```c++
void *get_ptr_node3d(Object *node) { return Object::cast_to<Node3D>(node); }
void *get_ptr_navagent(Object *node) { return Object::cast_to<NavigationAgent3D>(node); }
```

```
969ms
2256ms
966ms
2257ms
```