@harrisyu 2025-09-25
> > This pull cause both 3D and 2D collision shape not detect by body.
> 
> @harrisyu Although I've confirmed the bug with 3D, I haven't worked out yet how this could be affecting 2D collision shapes, do you have an MRP that shows this? (or even let me know if one of the demos isn't working correctly)

Sorry to see that late, I see first pull changed to 2D and 3D, but I only test 3D.
It works fine now with your later fix.